Add files via upload
Changed the nozzle throat diameter to get the peak mass flux below 2 since it is over that in the current sim and the pressure is pretty high. This is a starting point and we can tighten the nozzle throat for more impulse if needed but ideally we stay at around 800 psi average to avoid any possibility of another big boom.